requirements.txt
Flask
TensorFlow
Pillow

